Fix printout of Fragment type

+++ artdaq/ArtModules/TransferOutput_module.cc
@@ -915,7 +915,7 @@ art::TransferOutput::writeSubRun(SubRunPrincipal& srp)
 void
 art::TransferOutput::sendMessage_(uint64_t sequenceId, uint8_t messageType, TBufferFile& msg)
 {
-	TLOG(TLVL_DEBUG) << "Sending message with sequenceID=" << sequenceId << ", type=" << messageType << ", length=" << msg.Length();
+	TLOG(TLVL_DEBUG) << "Sending message with sequenceID=" << sequenceId << ", type=" << (int)messageType << ", length=" << msg.Length();
 	artdaq::NetMonHeader header;
 	header.data_length = static_cast<uint64_t>(msg.Length());
 	artdaq::Fragment
